core: move handling of hostname from plugins to core
How to write and read the machine hostname is something that has been
handled until now by plugins; this is questionable since the method
using for storing the hostname should depend only on the distro used
and not on which plugins are enabled.

This commit moves all hostname-related functions from plugins to the
core and allows to specify the method used to load and store the
hostname at build time with the

  --with-hostname-persist=default|suse|ifnet

configure option.

'default' method uses /etc/hostname to read and write the hostname. If
the ifcfg-rh plugin is enabled, the system tries to read the hostname
from /etc/sysconfig/network when it is not found in /etc/hostname and
both files are monitored to detect runtime changes.

When the selected method is 'suse', the hostname gets read from and
written to /etc/HOSTNAME; the file /etc/sysconfig/network/dhcp is also
read to detect if the hostname is dynamic and thus invalid. Both files
are monitored for changes.

'ifnet' method relies on /etc/conf.d/hostname for saving the hostname.